Moved package home page from https://github.com/iNamik/parser.go to https://github.com/iNamik/go_parser
It turns out that Go's clean functionality can be dangerous if you have a non-derived file with same exact name as the containing package folder.  This is currently a problem with the parser project with Go's clean removing the /parser.go/parser.go src file.

To reduce the potential for this problem in the future, I am moving all of my Go projects from '*.go' to 'go_*'

I also updated references to the iNamik go_container and go_lexer projects, which have already been renamed on github.